feat: implement custom hook for OAuth sign-in and refactor sign-in/sign-up pages fix: update userId reference in Stripe webhook handler fix: ensure user creation handles missing email in checkUser function docs: enhance Google OAuth setup instructions in documentation chore: improve error handling in test migration script
